Use GetFullPathNameA() to get absolute path of argv[0]
Character encoding on Windows is a real mess. For win32 API with 'A' post-fix, they only deal with ANSI string (not utf-8). So using std::filesystem here introduced locale bug. A windows version of GetFullPathNameA() will be better in the case, and it also helps to get rid of <filesystem> dependency.